clean up code & refactor
